# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Exam Preparation System (EPS) API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /programs:
    get:
      operationId: listPrograms
      tags: [programs]
      summary: List active programs (any authenticated user)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Program" }
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createProgram
      tags: [programs]
      summary: Create a new program (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProgramRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Program" }
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Program code already in use
...
  /exams/generate:
    post:
      operationId: generateExam
      tags: [exams]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateExamRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithQuestions"
        "400":
          $ref: "#/components/responses/BadRequest"
  /exams/{id}:
    get:
      operationId: getExam
      tags: [exams]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithQuestions"
        "404":
          $ref: "#/components/responses/NotFound"
  /exams/{id}/start:
    post:
      operationId: startExam
      tags: [exams]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamWithQuestions"
  /exams/{id}/submit:
    post:
      operationId: submitExam
      tags: [exams]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitExamRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /exams/{id}/review:
    get:
      operationId: getExamReview
      tags: [exams]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamReview"
  /users/{id}/exams:
    get:
      operationId: getUserExams
      tags: [users]
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Exam"
...
components:
  parameters:
    IdPath:
      in: path
      name: id
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Difficulty:
      type: string
      enum: [Easy, Medium, Hard]
...
    QuestionType:
      type: string
      enum: [single_choice, multiple_choice]
    ExamStatus:
      type: string
      enum: [generated, in_progress, submitted]
...
    Program:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        code: { type: string }
        status: { type: string }
      required: [id, name, code, status]
    CreateProgramRequest:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        code: { type: string, minLength: 1 }
        status: { type: string }
      required: [name, code]
...
    AnswerOption:
      type: object
      properties:
        id: { type: integer }
        questionId: { type: integer }
        answerText: { type: string }
        isCorrect: { type: boolean }
        displayOrder: { type: integer }
      required: [id, questionId, answerText, isCorrect, displayOrder]
...
    GenerateExamRequest:
      type: object
      properties:
        courseId: { type: integer }
        topicIds:
          type: array
          items: { type: integer }
        totalQuestions: { type: integer, minimum: 1, maximum: 100 }
        durationMinutes: { type: integer, nullable: true }
      required: [courseId, totalQuestions]
    ExamQuestionOption:
      type: object
      properties:
        id: { type: integer }
        answerText: { type: string }
      required: [id, answerText]
    ExamQuestion:
      type: object
      properties:
        id: { type: integer }
        questionId: { type: integer }
        title: { type: string }
        questionText: { type: string }
        questionType: { $ref: "#/components/schemas/QuestionType" }
        difficultyLevel: { $ref: "#/components/schemas/Difficulty" }
        topicName: { type: string, nullable: true }
        randomizedOrder: { type: integer }
        options:
          type: array
          items: { $ref: "#/components/schemas/ExamQuestionOption" }
        selectedAnswerOptionId: { type: integer, nullable: true }
      required:
        [id, questionId, title, questionText, questionType, difficultyLevel, randomizedOrder, options]
    Exam:
      type: object
      properties:
        id: { type: integer }
        userId: { type: integer }
        courseId: { type: integer }
        courseName: { type: string, nullable: true }
        examMode: { type: string }
        totalQuestions: { type: integer }
        durationMinutes: { type: integer, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        submittedAt: { type: string, format: date-time, nullable: true }
        score: { type: number, nullable: true }
        status: { $ref: "#/components/schemas/ExamStatus" }
        createdAt: { type: string, format: date-time }
      required: [id, userId, courseId, examMode, totalQuestions, status, createdAt]
    ExamWithQuestions:
      allOf:
        - $ref: "#/components/schemas/Exam"
        - type: object
          properties:
            questions:
              type: array
              items: { $ref: "#/components/schemas/ExamQuestion" }
          required: [questions]
    SubmitAnswer:
      type: object
      properties:
        examQuestionId: { type: integer }
        selectedAnswerOptionId: { type: integer, nullable: true }
      required: [examQuestionId]
    SubmitExamRequest:
      type: object
      properties:
        answers:
          type: array
          items: { $ref: "#/components/schemas/SubmitAnswer" }
      required: [answers]
    ExamResult:
      type: object
      properties:
        examId: { type: integer }
        score: { type: number }
        correctCount: { type: integer }
        totalQuestions: { type: integer }
        status: { $ref: "#/components/schemas/ExamStatus" }
      required: [examId, score, correctCount, totalQuestions, status]
    ReviewItem:
      type: object
      properties:
        examQuestionId: { type: integer }
        questionId: { type: integer }
        title: { type: string }
        questionText: { type: string }
        difficultyLevel: { $ref: "#/components/schemas/Difficulty" }
        topicName: { type: string, nullable: true }
        explanationText: { type: string, nullable: true }
        isCorrect: { type: boolean }
        selectedAnswerOptionId: { type: integer, nullable: true }
        correctAnswerOptionId: { type: integer, nullable: true }
        options:
          type: array
          items: { $ref: "#/components/schemas/AnswerOption" }
      required:
        [examQuestionId, questionId, title, questionText, difficultyLevel, isCorrect, options]
    ExamReview:
      type: object
      properties:
        exam: { $ref: "#/components/schemas/Exam" }
        items:
          type: array
          items: { $ref: "#/components/schemas/ReviewItem" }
      required: [exam, items]
...