# lib/api-zod/src/generated/types/createProgramRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exam Preparation System (EPS) API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateProgramRequest {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  code: string;
  status?: string;
}