# lib/api-zod/src/generated/types/program.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exam Preparation System (EPS) API
 * OpenAPI spec version: 0.1.0
 */

export interface Program {
  id: number;
  name: string;
  code: string;
  status: string;
}